Add visualize, update README, fix a few bugs in sarif.py.

+++ process_sarif/conformance.py
@@ -8,7 +8,8 @@ from process_sarif.sarif_helpers import get_sarif_in_build
 def main():
     log_path = "conformance_log.txt"
 
-    os.remove(log_path)
+    if os.path.exists(log_path):
+        os.remove(log_path)
 
     get_sarif_in_build(verbose=False, log_path=log_path)
